+++ src/App.js
@@ -10,6 +10,7 @@ import ProductDetails from "./pages/ProductDetails"
 import Navigation from "./components/Navigation"
 import Profilepage from "./pages/ProfilePage"
 import ShoppingCart from './pages/ShoppingCart';
+import CheckoutPage from "./pages/CheckoutPage"
 import { validateUser } from "./store/User/actions"
 import { Elements } from '@stripe/react-stripe-js';
 import { loadStripe } from "@stripe/stripe-js"
@@ -58,6 +59,10 @@ function App() {
           <Route
             path="/cartCheckout"
             component={ShoppingCart}
+          />
+          <Route
+            path="/purchase"
+            component={CheckoutPage}
           />
           </Elements>
         </Switch>
