+++ src/pages/CheckoutPage/index.js
@@ -1,0 +1,10 @@
+import React from "react"
+import PaymentInfo from "../../components/PaymentInfo"
+
+export default function CheckoutPage(){
+    return (
+        <div>
+            <PaymentInfo />
+        </div>
+    )
+}
